PLANNER-401: Moved StatsUtil, fixed failure handling and implement interface correctly
