+++ optaplanner-benchmark/src/test/java/org/optaplanner/benchmark/impl/ranking/AbstractSolverRankingComparatorTest.java
@@ -41,6 +41,7 @@ public abstract class AbstractSolverRankingComparatorTest {
         SimpleScore scoreObject = SimpleScore.valueOf(score);
         SimpleScore bestScoreObject = SimpleScore.valueOf(bestScore);
         SimpleScore worstScoreObject = SimpleScore.valueOf(worstScore);
+        singleBenchmarkResult.setFailureCount(0);
         singleBenchmarkResult.setAverageScore(scoreObject);
         singleBenchmarkResult.setWinningScoreDifference(scoreObject.subtract(bestScoreObject));
         singleBenchmarkResult.setWorstScoreDifferencePercentage(
@@ -56,6 +57,7 @@ public abstract class AbstractSolverRankingComparatorTest {
         HardSoftLongScore scoreObject = HardSoftLongScore.valueOf(hardScore, softScore);
         HardSoftLongScore bestScoreObject = HardSoftLongScore.valueOf(hardBestScore, softBestScore);
         HardSoftLongScore worstScoreObject = HardSoftLongScore.valueOf(hardWorstScore, softWorstScore);
+        singleBenchmarkResult.setFailureCount(0);
         singleBenchmarkResult.setAverageScore(scoreObject);
         singleBenchmarkResult.setWinningScoreDifference(scoreObject.subtract(bestScoreObject));
         singleBenchmarkResult.setWorstScoreDifferencePercentage(
